+++ tools/lib/lmp.h
@@ -14,8 +14,8 @@
 
 /* LMP file structures */
 typedef struct {
-    uint16_t width;
-    uint16_t height;
+    uint32_t width;
+    uint32_t height;
     uint8_t *data;
     size_t data_size;
 } lmp_file_t;
